+++ crates/krilla/src/font/colr.rs
@@ -458,119 +458,3 @@ impl ColorPainter for ColrBuilder {
         stack.push(Instruction::Layer(blend, instructions));
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use krilla_macros::visreg;
-    use skrifa::GlyphId;
-    use tiny_skia_path::Point;
-
-    use crate::document::Document;
-    use crate::font::Font;
-    use crate::path::{Fill, Stroke};
-    use crate::surface::{Surface, TextDirection};
-    use crate::tests::{
-        all_glyphs_to_pdf, blue_stroke, purple_fill, COLR_TEST_GLYPHS, NOTO_COLOR_EMOJI_COLR,
-    };
-
-    #[visreg(document)]
-    fn colr_test_glyphs(document: &mut Document) {
-        let font_data = COLR_TEST_GLYPHS.clone();
-
-        let glyphs = (0..=220)
-            .map(|n| (GlyphId::new(n), "".to_string()))
-            .collect::<Vec<_>>();
-
-        all_glyphs_to_pdf(font_data, Some(glyphs), false, true, document);
-    }
-
-    #[visreg]
-    fn colr_context_color(surface: &mut Surface) {
-        let font_data = COLR_TEST_GLYPHS.clone();
-        let font = Font::new(font_data, 0, true).unwrap();
-
-        let text = [
-            0xf0b00, 0xf0b01, 0xf0b02, 0xf0b03, 0xf0b04, 0xf0b05, 0xf0b06, 0xf0b07,
-        ]
-        .into_iter()
-        .map(|n| char::from_u32(n).unwrap().to_string())
-        .collect::<Vec<_>>()
-        .join(" ");
-
-        surface.fill_text(
-            Point::from_xy(0., 30.0),
-            Fill::default(),
-            font.clone(),
-            15.0,
-            &[],
-            &text,
-            false,
-            TextDirection::Auto,
-        );
-
-        surface.fill_text(
-            Point::from_xy(0., 50.0),
-            purple_fill(1.0),
-            font.clone(),
-            15.0,
-            &[],
-            &text,
-            false,
-            TextDirection::Auto,
-        );
-
-        surface.fill_text(
-            Point::from_xy(0., 70.0),
-            purple_fill(1.0),
-            font.clone(),
-            15.0,
-            &[],
-            &text,
-            true,
-            TextDirection::Auto,
-        );
-
-        surface.stroke_text(
-            Point::from_xy(0., 130.0),
-            Stroke::default(),
-            font.clone(),
-            15.0,
-            &[],
-            &text,
-            false,
-            TextDirection::Auto,
-        );
-
-        // Since it a COLR glyph, it will still be filled, but the color should be taken from
-        // the stroke.
-        surface.stroke_text(
-            Point::from_xy(0., 150.0),
-            blue_stroke(1.0),
-            font.clone(),
-            15.0,
-            &[],
-            &text,
-            false,
-            TextDirection::Auto,
-        );
-
-        surface.stroke_text(
-            Point::from_xy(0., 170.0),
-            blue_stroke(1.0),
-            font.clone(),
-            15.0,
-            &[],
-            &text,
-            true,
-            TextDirection::Auto,
-        );
-    }
-
-    // We don't run on pdf.js because it leads to a high pixel difference in CI
-    // for some reason.
-    #[visreg(document, pdfium, mupdf, pdfbox, ghostscript, poppler, quartz)]
-    fn noto_color_emoji_colr(document: &mut Document) {
-        let font_data = NOTO_COLOR_EMOJI_COLR.clone();
-        all_glyphs_to_pdf(font_data, None, false, true, document);
-    }
-}
